Add logout functionality refactor
